Move ToValue and FromValue to internal.
