+++ tests/EndToEndTests.hs
@@ -14,8 +14,7 @@ import GraphQL (makeSchema, compileQuery, executeQuery, interpretAnonymousQuery,
 import GraphQL.API (Object, Field)
 import GraphQL.Internal.Syntax.AST (Variable(..))
 import GraphQL.Resolver ((:<>)(..), Handler)
-import GraphQL.Value (makeName)
-import GraphQL.Value.ToValue (ToValue(..))
+import GraphQL.Value (ToValue(..), makeName)
 import Test.Tasty (TestTree)
 import Test.Tasty.Hspec (testSpec, describe, it, shouldBe)
 import Text.RawString.QQ (r)
